# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Meditiya Sathi - Digital Society Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notices:
    get:
      operationId: listNotices
      tags: [notices]
      summary: List notices
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: pinned
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of notices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notice"
    post:
      operationId: createNotice
      tags: [notices]
      summary: Create notice (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoticeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notice"
  /notices/{id}:
    get:
      operationId: getNotice
      tags: [notices]
      summary: Get notice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notice"
    patch:
      operationId: updateNotice
      tags: [notices]
      summary: Update notice (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoticeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notice"
    delete:
      operationId: deleteNotice
      tags: [notices]
      summary: Delete notice (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── GALLERY ────────────────────────────────────────────────────────────────
  /gallery/albums:
    get:
      operationId: listAlbums
      tags: [gallery]
      summary: List gallery albums
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: festival
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Albums
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Album"
    post:
      operationId: createAlbum
      tags: [gallery]
      summary: Create album (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlbumInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
  /gallery/albums/{id}/photos:
    get:
      operationId: listAlbumPhotos
      tags: [gallery]
      summary: List photos in album
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Photos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryPhoto"
    post:
      operationId: addPhotoToAlbum
      tags: [gallery]
      summary: Add photo to album (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryPhotoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryPhoto"
...
  /volunteers:
    get:
      operationId: listVolunteers
      tags: [volunteers]
      summary: List volunteers
      parameters:
        - name: festival
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Volunteers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Volunteer"
    post:
      operationId: joinVolunteer
      tags: [volunteers]
      summary: Apply to become a volunteer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VolunteerInput"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Volunteer"

  # ── COMPETITIONS ───────────────────────────────────────────────────────────
...
  /services/requests:
    get:
      operationId: listServiceRequests
      tags: [services]
      summary: List service requests (admin or own)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, resolved, all]
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Service requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceRequest"
    post:
      operationId: createServiceRequest
      tags: [services]
      summary: Submit a service request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceRequestInput"
      responses:
        "201":
          description: Request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
  /services/requests/{id}:
    patch:
      operationId: updateServiceRequest
      tags: [services]
      summary: Update service request status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceRequestUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"

  # ── MARKETPLACE ────────────────────────────────────────────────────────────
  /marketplace:
    get:
      operationId: listMarketplaceItems
      tags: [marketplace]
      summary: List marketplace listings
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [sell, buy, donate, all]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketplaceItem"
    post:
      operationId: createMarketplaceItem
      tags: [marketplace]
      summary: Create listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketplaceItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketplaceItem"
  /marketplace/{id}:
    delete:
      operationId: deleteMarketplaceItem
      tags: [marketplace]
      summary: Delete listing (owner/admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── LOST & FOUND ───────────────────────────────────────────────────────────
  /lostfound:
    get:
      operationId: listLostFoundItems
      tags: [lostfound]
      summary: List lost and found items
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [lost, found, all]
        - name: resolved
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LostFoundItem"
    post:
      operationId: createLostFoundItem
      tags: [lostfound]
      summary: Report lost or found item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LostFoundItemInput"
      responses:
        "201":
          description: Reported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LostFoundItem"

  # ── COMMITTEE ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Notice:
      type: object
      required: [id, title, content, category, isPinned, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        category:
          type: string
        isPinned:
          type: boolean
        attachmentUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    NoticeInput:
      type: object
      required: [title, content, category]
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
        isPinned:
          type: boolean
        attachmentUrl:
          type: ["string", "null"]
    NoticeUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
        isPinned:
          type: boolean
        attachmentUrl:
          type: ["string", "null"]

    # ── GALLERY ─────────────────────────────────────────────────────────────
    Album:
      type: object
      required: [id, title, year, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        year:
          type: integer
        festival:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
        photoCount:
          type: integer
        createdAt:
          type: string
    AlbumInput:
      type: object
      required: [title, year]
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        year:
          type: integer
        festival:
          type: ["string", "null"]
        coverImageUrl:
          type: ["string", "null"]
    GalleryPhoto:
      type: object
      required: [id, albumId, imageUrl, createdAt]
      properties:
        id:
          type: integer
        albumId:
          type: integer
        imageUrl:
          type: string
        caption:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        createdAt:
          type: string
    GalleryPhotoInput:
      type: object
      required: [imageUrl]
      properties:
        imageUrl:
          type: string
        caption:
          type: ["string", "null"]
        isFeatured:
          type: boolean

    # ── DONATIONS ───────────────────────────────────────────────────────────
...
    Volunteer:
      type: object
      required: [id, name, phone, role, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
        flatNumber:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        festivalId:
          type: ["integer", "null"]
        userId:
          type: ["string", "null"]
        createdAt:
          type: string
    VolunteerInput:
      type: object
      required: [name, phone, role]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
        flatNumber:
          type: ["string", "null"]
        festivalId:
          type: ["integer", "null"]

    # ── COMPETITIONS ────────────────────────────────────────────────────────
...
    ServiceRequest:
      type: object
      required: [id, category, description, status, requesterName, createdAt]
      properties:
        id:
          type: integer
        category:
          type: string
          enum: [electrician, plumber, cleaning, maintenance, water_issue, street_light, security, other]
        description:
          type: string
        status:
          type: string
          enum: [pending, in_progress, resolved]
        requesterName:
          type: string
        requesterPhone:
          type: string
        flatNumber:
          type: ["string", "null"]
        userId:
          type: ["string", "null"]
        adminNote:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ServiceRequestInput:
      type: object
      required: [category, description, requesterName, requesterPhone]
      properties:
        category:
          type: string
        description:
          type: string
        requesterName:
          type: string
        requesterPhone:
          type: string
        flatNumber:
          type: ["string", "null"]
    ServiceRequestUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_progress, resolved]
        adminNote:
          type: ["string", "null"]

    # ── MARKETPLACE ─────────────────────────────────────────────────────────
    MarketplaceItem:
      type: object
      required: [id, title, category, listingType, sellerName, sellerPhone, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          enum: [furniture, books, electronics, cycles, other]
        listingType:
          type: string
          enum: [sell, buy, donate]
        price:
          type: ["number", "null"]
        imageUrl:
          type: ["string", "null"]
        sellerName:
          type: string
        sellerPhone:
          type: string
        userId:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        createdAt:
          type: string
    MarketplaceItemInput:
      type: object
      required: [title, category, listingType, sellerName, sellerPhone]
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        listingType:
          type: string
        price:
          type: ["number", "null"]
        imageUrl:
          type: ["string", "null"]
        sellerName:
          type: string
        sellerPhone:
          type: string

    # ── LOST & FOUND ────────────────────────────────────────────────────────
    LostFoundItem:
      type: object
      required: [id, type, title, description, contactName, contactPhone, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [lost, found]
        title:
          type: string
        description:
          type: string
        location:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        contactName:
          type: string
        contactPhone:
          type: string
        isResolved:
          type: boolean
        userId:
          type: ["string", "null"]
        createdAt:
          type: string
    LostFoundItemInput:
      type: object
      required: [type, title, description, contactName, contactPhone]
      properties:
        type:
          type: string
          enum: [lost, found]
        title:
          type: string
        description:
          type: string
        location:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        contactName:
          type: string
        contactPhone:
          type: string

    # ── COMMITTEE ───────────────────────────────────────────────────────────
...